requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.1.0
blinker==1.9.0
Brotli==1.1.0
build==0.10.0
CacheControl==0.12.14
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudinary==1.44.1
ConfigArgParse==1.7
contourpy==1.3.3
crashtest==0.4.1
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
dulwich==0.21.7
fastapi==0.110.0
sqlalchemy==2.0.43
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
fonttools==4.59.1
gevent==24.11.1
geventhttpclient==2.3.3
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.1.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.27.2
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.13.1
kiwisolver==1.4.9
lockfile==0.12.2
locust==2.33.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
narwhals==2.1.2
nltk==3.9.1
numpy==1.26.4
openai==1.14.2
packaging==23.2
pandas==2.2.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.12.0
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyzmq==26.2.1
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==0.10.1
rich==13.9.4
rpds-py==0.22.3
seaborn==0.13.2
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.36.3
streamlit==1.32.2
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.2
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.29.0
virtualenv==20.21.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
xattr==0.10.1
zope.event==5.0
zope.interface==7.2
